requirements.txt
Django==5.2.5
mysqlclient==2.2.7
python-decouple==3.8
cloudinary==1.34.0
django-cloudinary-storage==0.3.0
requests==2.31.0
djangorestframework==3.20.0
django-widget-tweaks==1.4.13
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==7.1.1
djongo==2.4.1
pymongo[srv]==4.7.2
